fix incorrect behavior of switch statement
switch statement in js works as follows: 1st we need to search for the
case matching given value. - we do this in 1st loop, also remembering
position of "default". 2nd thing is to execute all the code at matching
case (if any) and all the code below chat matching case (until we find
"break").